last update

+++ CPP09/ex01/RPN.cpp
@@ -22,17 +22,17 @@ static bool isOperand(std::string& c) {
 }
 
 static int executeOperand(std::string& token, int &a, int& b) {
-	double	res;
+	long	res;
 	if (token == "+")
-		res = a + b;
+		res = (long)a + b;
 	else if (token == "-")
-		res = b - a;
+		res = (long)b - a;
 	else if (token == "*" )
-		res = b * a;
+		res = (long)b * a;
 	else if (token == "/") {
 		if (a == 0)
 			throw RPN::Error();
-		res = b / a;
+		res = (long)b / a;
 	}
 	if (res > std::numeric_limits<int>::max() 
 		|| res < std::numeric_limits<int>::min()) {
